MOT new features: 1. JIT support for stored procedures 2. MVCC & Cross-Txn support 3. Alter Table support (Add/Drop/Rename Column) 4. Low RTO 5. More stabilizations

+++ src/gausskernel/storage/mot/core/infra/stats/level_statistic_variable.cpp
@@ -42,9 +42,9 @@ void LevelStatisticVariable::Summarize(bool updateTstamp)
 void LevelStatisticVariable::Print(LogLevel logLevel) const
 {
     MOT_LOG(logLevel,
-        "%s={ current level: %" PRId64 " %s, peak: %" PRIu64 " %s [%" PRIu64 " samples] }",
+        "%s={ current level: %" PRIu64 " %s, peak: %" PRIu64 " %s [%" PRIu64 " samples] }",
         m_name,
-        m_levelSaved / m_factor,
+        static_cast<uint64_t>(m_levelSaved) / m_factor,
         m_units,
         m_peakSaved / m_factor,
         m_units,
